requirements.txt
# Core framework and async support
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# Database and caching
asyncpg==0.29.0
redis==5.0.1
redis[hiredis]==5.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
supabase==2.3.4
greenlet==3.0.3

# Security and encryption
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# HTTP client and utilities
httpx==0.25.2
python-dotenv==1.0.1

# AI vendor SDKs
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2

# Testing and development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
streamlit==1.31.1 